+++ src/ocean/testbed/mo_ocean_testbed_zstar.f90
@@ -54,7 +54,7 @@ MODULE mo_ocean_testbed_zstar
     & timer_extra20, timer_extra21, timer_normal_veloc, timer_scalar_prod_veloc
 
   USE mo_dynamics_config,           ONLY: nold, nnew
-  USE mo_physical_constants,        ONLY: grav, clw, rho_ref, Tf
+  USE mo_physical_constants,        ONLY: grav, clw, rho_ref
   USE mo_ocean_initialization,      ONLY: is_initial_timestep
   USE mo_ocean_types, ONLY: t_hydro_ocean_state
   USE mo_ocean_time_events,   ONLY: ocean_time_nextStep, isCheckpoint, isEndOfThisRun, newNullDatetime