+++ src/sea_ice/thermo/mo_ice_fast.f90
@@ -27,8 +27,8 @@ MODULE mo_ice_fast
 
   USE mo_timer,               ONLY: timer_start, timer_stop, timer_ice_fast
 
-  USE mo_physical_constants,  ONLY: rhoi, ki, Tf, ci
-  USE mo_sea_ice_nml,         ONLY: i_ice_therm, hci_layer
+  USE mo_physical_constants,  ONLY: rhoi, ki, ci
+  USE mo_sea_ice_nml,         ONLY: i_ice_therm, hci_layer, Tf
 
   USE mo_ice_winton,          ONLY: set_ice_temp_winton
   USE mo_ice_zerolayer,       ONLY: set_ice_temp_zerolayer, set_ice_temp_zerolayer_analytical